# swagger/api/v1/manifest.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api/v1/manifest.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "CloudEventsService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/cloudevents": {
      "post": {
        "operationId": "CloudEventsService_Send",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CloudEventSendResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "description": "CloudEvent is copied from\nhttps://github.com/cloudevents/spec/blob/master/protobuf-format.md.",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/protobufCloudEvent"
            }
          }
        ],
        "tags": [
          "CloudEventsService"
        ]
      }
    },
    "/v1/resources/{id}/status": {
      "get": {
        "operationId": "CloudEventsService_Watch",
        "responses": {
          "200": {
            "description": "A successful response.(streaming responses)",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "$ref": "#/definitions/protobufCloudEvent"
                },
                "error": {
                  "$ref": "#/definitions/googlerpcStatus"
                }
              },
              "title": "Stream result of protobufCloudEvent"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "CloudEventsService"
        ]
      }
    }
  },
  "definitions": {
    "googlerpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string",
          "description": "A URL/resource name that uniquely identifies the type of the serialized\nprotocol buffer message. This string must contain at least\none \"/\" character. The last segment of the URL's path must represent\nthe fully qualified name of the type (as in\n`path/google.protobuf.Duration`). The name should be in a canonical form\n(e.g., leading \".\" is not accepted).\n\nIn practice, teams usually precompile into the binary all types that they\nexpect it to use in the context of Any. However, for URLs which use the\nscheme `http`, `https`, or no scheme, one can optionally set up a type\nserver that maps type URLs to message definitions as follows:\n\n* If no scheme is provided, `https` is assumed.\n* An HTTP GET on the URL must yield a [google.protobuf.Type][]\n  value in binary format, or produce an error.\n* Applications are allowed to cache lookup results based on the\n  URL, or have them precompiled into a binary to avoid any\n  lookup. Therefore, binary compatibility needs to be preserved\n  on changes to types. (Use versioned type names to manage\n  breaking changes.)\n\nNote: this functionality is not currently available in the official\nprotobuf release, and it is not used for type URLs beginning with\ntype.googleapis.com.\n\nSchemes other than `http`, `https` (or the empty scheme) might be\nused with implementation specific semantics."
        }
      },
      "additionalProperties": {},
      "description": "`Any` contains an arbitrary serialized protocol buffer message along with a\nURL that describes the type of the serialized message.\n\nProtobuf library provides support to pack/unpack Any values in the form\nof utility functions or additional generated methods of the Any type.\n\nExample 1: Pack and unpack a message in C++.\n\n    Foo foo = ...;\n    Any any;\n    any.PackFrom(foo);\n    ...\n    if (any.UnpackTo(\u0026foo)) {\n      ...\n    }\n\nExample 2: Pack and unpack a message in Java.\n\n    Foo foo = ...;\n    Any any = Any.pack(foo);\n    ...\n    if (any.is(Foo.class)) {\n      foo = any.unpack(Foo.class);\n    }\n    // or ...\n    if (any.isSameTypeAs(Foo.getDefaultInstance())) {\n      foo = any.unpack(Foo.getDefaultInstance());\n    }\n\nExample 3: Pack and unpack a message in Python.\n\n    foo = Foo(...)\n    any = Any()\n    any.Pack(foo)\n    ...\n    if any.Is(Foo.DESCRIPTOR):\n      any.Unpack(foo)\n      ...\n\nExample 4: Pack and unpack a message in Go\n\n     foo := \u0026pb.Foo{...}\n     any, err := anypb.New(foo)\n     if err != nil {\n       ...\n     }\n     ...\n     foo := \u0026pb.Foo{}\n     if err := any.UnmarshalTo(foo); err != nil {\n       ...\n     }\n\nThe pack methods provided by protobuf library will by default use\n'type.googleapis.com/full.type.name' as the type URL and the unpack\nmethods only use the fully qualified type name after the last '/'\nin the type URL, for example \"foo.bar.com/x/y.z\" will yield type\nname \"y.z\".\n\nJSON\n\nThe JSON representation of an `Any` value uses the regular\nrepresentation of the deserialized, embedded message, with an\nadditional field `@type` which contains the type URL. Example:\n\n    package google.profile;\n    message Person {\n      string first_name = 1;\n      string last_name = 2;\n    }\n\n    {\n      \"@type\": \"type.googleapis.com/google.profile.Person\",\n      \"firstName\": \u003cstring\u003e,\n      \"lastName\": \u003cstring\u003e\n    }\n\nIf the embedded message type is well-known and has a custom JSON\nrepresentation, that representation will be embedded adding a field\n`value` which holds the custom JSON in addition to the `@type`\nfield. Example (for message [google.protobuf.Duration][]):\n\n    {\n      \"@type\": \"type.googleapis.com/google.protobuf.Duration\",\n      \"value\": \"1.212s\"\n    }"
    },
    "protobufCloudEvent": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "description": "Unique event identifier."
        },
        "source": {
          "type": "string",
          "description": "URI of the event source."
        },
        "specVersion": {
          "type": "string",
          "description": "Version of the spec in use."
        },
        "type": {
          "type": "string",
          "description": "Event type identifier."
        },
        "attributes": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/protobufCloudEventAttributeValue"
          },
          "title": "Optional \u0026 Extension Attributes"
        },
        "binaryData": {
          "type": "string",
          "format": "byte",
          "description": "If the event is binary data then the datacontenttype attribute\nshould be set to an appropriate media-type."
        },
        "textData": {
          "type": "string",
          "description": "If the event is string data then the datacontenttype attribute\nshould be set to an appropriate media-type such as application/json."
        },
        "protoData": {
          "$ref": "#/definitions/protobufAny",
          "description": "If the event is a protobuf then it must be encoded using this Any\ntype. The datacontenttype attribute should be set to\napplication/protobuf and the dataschema attribute set to the message\ntype."
        }
      },
      "description": "CloudEvent is copied from\nhttps://github.com/cloudevents/spec/blob/master/protobuf-format.md."
    },
    "protobufCloudEventAttributeValue": {
      "type": "object",
      "properties": {
        "ceBoolean": {
          "type": "boolean",
          "description": "Boolean value."
        },
        "ceInteger": {
          "type": "integer",
          "format": "int32",
          "description": "Integer value."
        },
        "ceString": {
          "type": "string",
          "description": "String value."
        },
        "ceBytes": {
          "type": "string",
          "format": "byte",
          "description": "Byte string value."
        },
        "ceUri": {
          "type": "string",
          "description": "URI value."
        },
        "ceUriRef": {
          "type": "string",
          "description": "URI reference value."
        },
        "ceTimestamp": {
          "type": "string",
          "format": "date-time",
          "description": "Timestamp value."
        }
      },
      "description": "CloudEventAttribute enables extensions to use any of the seven allowed\ndata types as the value of an envelope key."
    },
    "v1CloudEventSendResponse": {
      "type": "object",
      "properties": {
        "status": {
          "$ref": "#/definitions/v1CloudEventSendResponseStatus"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "v1CloudEventSendResponseStatus": {
      "type": "string",
      "enum": [
        "OK",
        "ERROR"
      ],
      "default": "OK"
    }
  }
}